fix: hook for pagination in new file
